Update types to better match reality.
This also fixes an issue I discovered while figuring out what type `indentExclusionsRanges` was supposed to be. `MagicString#clone` does not appear to work when the object was constructed with a non-tuple. It seemed to expect an array of nodes – or something with `start` and `end` properties – but generated an array of tuples. Calling `#clone` multiple times no matter what the starting value would have resulted in an incorrect clone. I'm not 100% sure that I picked the right type, but this seemed the most consistent choice.