Don't use `declare module` for TypeScript types
As I understand things, the `declare module` TypeScript syntax creates a
globally-importable name and says that any import with that name maps to the
given type. This means it usually works for libraries, but breaks down when
you have multiple versions of the same library loaded, since they each redefine
the same global name. It's really meant for things like Node.js built-in
modules. Instead, `.d.ts` files in libraries should just do exports at the
top-level, and TypeScript allows different versions of the same library to
peacefully coexist in that case.

I ran into a dependency hell problem that I think was caused by this issue where
decaffeinate pulled in three different versions of magic-string that conflicted.
I was able to change the dependency tree so it's just one version, which works
in the near term, but this change should prevent that from happening again.