Support a function for includeContent in Bundle
Suppose you have a `magic-string` built up from a variety of sources, some of which can be loaded by the runtime and some that can't. Those that can be loaded might be, for example, public http(s) urls. In this scenario, you might want to embed source content for those sources that _can't_ be loaded at runtime.