Reduce memory overhead of sourcemapLocations
It was using an object to store positions in the string. This meant that
those offsets were first stringified and then added as object
properties. For a 200K JS code file (as parsed by rollup), the
sourcemapLocations object was taking up 980K. Switching it to a BitSet
that uses one bit per offset makes it take up 36K instead.

Fixes Rich-Harris/magic-string#167