Add tests for the copy feature
In test/dev builds, DEBUG is replaced with true, which runs additional
code marking the 'duplicate' chunks created by .copy() with chunk.isCopy.

This is detected in integrity checker in tests - these chunks are,
as described in previous commit, not included in the indexes, which
isn't easy to solve, so these chunks are just excluded from the check.

In production code (verified with npm run publish), DEBUG is replaced
with false and the code becomes unreachable and is tree-shaked.