feat(x_google_ignoreList): initial support for ignore lists
This patch introduces support for the `x_google_ignoreList` field
to the `SourceMap` class. This extension was added to the Source Map
Revision 3 Proposal[^1] to allow build tools to provide hints to
debuggers (e.g. browser DevTools) about which files contain library
or framework, and should thus be ignored by default for the purpose
of stepping, break on exceptions, and the like[^2].

With this change it's possible for consumers of magic-string's
`SourceMap` class to thread through and serialize the newly added
`x_google_ignoreList` field (for example this is needed for rollup
to support `x_google_ignoreList`).

In a follow up change, we will also add support to the `Bundle`
class to mark certain sources as ignore listed.

[^1]: https://docs.google.com/document/d/1U1RGAehQwRypUTovF1KRlpiOFze0b-_2gc6fAH0KY0k
[^2]: https://developer.chrome.com/blog/devtools-better-angular-debugging/